Move time-related instances to trait
