+++ modules/java8/src/main/scala/io/circe/java8/time/package.scala
@@ -1,106 +1,3 @@
 package io.circe.java8
 
-import io.circe.{ Decoder, DecodingFailure, Encoder, Json }
-import java.time.{ Instant, LocalDate, LocalDateTime, LocalTime, OffsetDateTime, ZonedDateTime }
-import java.time.format.{ DateTimeFormatter, DateTimeParseException }
-import java.time.format.DateTimeFormatter.{
-  ISO_LOCAL_DATE,
-  ISO_LOCAL_DATE_TIME,
-  ISO_LOCAL_TIME,
-  ISO_OFFSET_DATE_TIME,
-  ISO_ZONED_DATE_TIME
-}
-
-package object time {
-  implicit final val decodeInstant: Decoder[Instant] =
-    Decoder.instance { c =>
-      c.as[String] match {
-        case Right(s) => try Right(Instant.parse(s)) catch {
-          case _: DateTimeParseException => Left(DecodingFailure("Instant", c.history))
-        }
-        case l @ Left(_) => l.asInstanceOf[Decoder.Result[Instant]]
-      }
-    }
-
-  implicit final val encodeInstant: Encoder[Instant] = Encoder.instance(time => Json.fromString(time.toString))
-
-  final def decodeLocalDateTime(formatter: DateTimeFormatter): Decoder[LocalDateTime] =
-    Decoder.instance { c =>
-      c.as[String] match {
-        case Right(s) => try Right(LocalDateTime.parse(s, formatter)) catch {
-          case _: DateTimeParseException => Left(DecodingFailure("LocalDateTime", c.history))
-        }
-        case l @ Left(_) => l.asInstanceOf[Decoder.Result[LocalDateTime]]
-      }
-    }
-
-  final def encodeLocalDateTime(formatter: DateTimeFormatter): Encoder[LocalDateTime] =
-    Encoder.instance(time => Json.fromString(time.format(formatter)))
-
-  implicit final val decodeLocalDateTimeDefault: Decoder[LocalDateTime] = decodeLocalDateTime(ISO_LOCAL_DATE_TIME)
-  implicit final val encodeLocalDateTimeDefault: Encoder[LocalDateTime] = encodeLocalDateTime(ISO_LOCAL_DATE_TIME)
-
-  final def decodeZonedDateTime(formatter: DateTimeFormatter): Decoder[ZonedDateTime] =
-    Decoder.instance { c =>
-      c.as[String] match {
-        case Right(s) => try Right(ZonedDateTime.parse(s, formatter)) catch {
-          case _: DateTimeParseException => Left(DecodingFailure("ZonedDateTime", c.history))
-        }
-        case l @ Left(_) => l.asInstanceOf[Decoder.Result[ZonedDateTime]]
-      }
-    }
-
-  final def encodeZonedDateTime(formatter: DateTimeFormatter): Encoder[ZonedDateTime] =
-    Encoder.instance(time => Json.fromString(time.format(formatter)))
-
-  implicit final val decodeZonedDateTimeDefault: Decoder[ZonedDateTime] = decodeZonedDateTime(ISO_ZONED_DATE_TIME)
-  implicit final val encodeZonedDateTimeDefault: Encoder[ZonedDateTime] = encodeZonedDateTime(ISO_ZONED_DATE_TIME)
-
-  final def decodeOffsetDateTime(formatter: DateTimeFormatter): Decoder[OffsetDateTime] =
-    Decoder.instance { c =>
-      c.as[String] match {
-        case Right(s) => try Right(OffsetDateTime.parse(s, formatter)) catch {
-          case _: DateTimeParseException => Left(DecodingFailure("OffsetDateTime", c.history))
-        }
-        case l @ Left(_) => l.asInstanceOf[Decoder.Result[OffsetDateTime]]
-      }
-    }
-
-  final def encodeOffsetDateTime(formatter: DateTimeFormatter): Encoder[OffsetDateTime] =
-    Encoder.instance(time => Json.fromString(time.format(formatter)))
-
-  implicit final val decodeOffsetDateTimeDefault: Decoder[OffsetDateTime] = decodeOffsetDateTime(ISO_OFFSET_DATE_TIME)
-  implicit final val encodeOffsetDateTimeDefault: Encoder[OffsetDateTime] = encodeOffsetDateTime(ISO_OFFSET_DATE_TIME)
-
-  final def decodeLocalDate(formatter: DateTimeFormatter): Decoder[LocalDate] =
-    Decoder.instance { c =>
-      c.as[String] match {
-        case Right(s) => try Right(LocalDate.parse(s, formatter)) catch {
-          case _: DateTimeParseException => Left(DecodingFailure("LocalDate", c.history))
-        }
-        case l @ Left(_) => l.asInstanceOf[Decoder.Result[LocalDate]]
-      }
-    }
-
-  final def encodeLocalDate(formatter: DateTimeFormatter): Encoder[LocalDate] =
-    Encoder.instance(time => Json.fromString(time.format(formatter)))
-
-  implicit final val decodeLocalDateDefault: Decoder[LocalDate] = decodeLocalDate(ISO_LOCAL_DATE)
-  implicit final val encodeLocalDateDefault: Encoder[LocalDate] = encodeLocalDate(ISO_LOCAL_DATE)
-
-  final def decodeLocalTime(formatter: DateTimeFormatter): Decoder[LocalTime] =
-    Decoder.instance { c =>
-      c.as[String] match {
-        case Right(s) => try Right(LocalTime.parse(s, formatter)) catch {
-          case _: DateTimeParseException => Left(DecodingFailure("LocalTime", c.history))
-        }
-        case l @ Left(_) => l.asInstanceOf[Decoder.Result[LocalTime]]
-      }
-    }
-
-  final def encodeLocalTime(formatter: DateTimeFormatter): Encoder[LocalTime] =
-    Encoder.instance(time => Json.fromString(time.format(formatter)))
-
-  implicit final val decodeLocalTimeDefault: Decoder[LocalTime] = decodeLocalTime(ISO_LOCAL_TIME)
-  implicit final val encodeLocalTimeDefault: Encoder[LocalTime] = encodeLocalTime(ISO_LOCAL_TIME)
-}
+package object time extends TimeInstances
